Improve match logging
I've switched to using waldo here in order to generate minimal diffs in the case of a failed match. I'm not committed to this precise log format, but this is where I ended up with.

Long run, I think it would be useful to include the index of the recorded request in the log message, but that will require some `HTTPInteractionList` refactoring first.