Don't document/export internal functions/classes
I left the Cassette class exported/documented for now, since it is technically part of the exported interface of the package because it's returned by `use_cassette()` etc. I could be persauded that it should be removed too (but that should wait until after #429 to avoid a bunch of merge conflicts)

Fixes #374.