Implement `insert_example_cassette()`
I thought originally that this would just be a documentation issue, but as I got deeper into it I realised that there's enough meat here to be worth a helper. Examples store cassettes in`inst/_vcr` so they're available once the package is installed.

I'm not sure how to test this, but I've set the default for `record` in such a way that the example should error during `R CMD check` if I've messed the code.

Fixes #309